Add method preventDefault()
React uses synthetic events to handle events from button, input and form elements. ... It demonstrates how to add an item to a list by using a form element with input and button elements. In this case, a preventDefault is called on the event when submitting the form to prevent a browser reload/refresh.